move createObservablePart into its own file
