+++ src/core/observable-api/unique-behavior-subject.ts
@@ -1,4 +1,4 @@
-import { BehaviorSubject, distinctUntilChanged, map, Observable, shareReplay } from "rxjs";
+import { BehaviorSubject } from "rxjs";
 
 
 // TODO: Should this handle array as well?
@@ -58,38 +58,15 @@ export function appendToFrozenArray<T>(data: T[], entry: T, uniqueMethod?: (exis
 
 
 
-type MappingFunction<T, R> = (mappable: T) => R;
-type MemoizationFunction<R> = (previousResult: R, currentResult: R) => boolean;
+export type MappingFunction<T, R> = (mappable: T) => R;
+export type MemoizationFunction<R> = (previousResult: R, currentResult: R) => boolean;
 
-function defaultMemoization(previousValue: any, currentValue: any): boolean {
+export function defaultMemoization(previousValue: any, currentValue: any): boolean {
 	if (typeof previousValue === 'object' && typeof currentValue === 'object') {
 	return naiveObjectComparison(previousValue, currentValue);
 	}
 	return previousValue === currentValue;
 }
-
-/**
- * @export
- * @method createObservablePart
- * @param {Observable<T>} source - RxJS Subject to use for this Observable.
- * @param {(mappable: T) => R} mappingFunction - Method to return the part for this Observable to return.
- * @param {(previousResult: R, currentResult: R) => boolean} [memoizationFunction] - Method to Compare if the data has changed. Should return true when data is different.
- * @description - Creates a RxJS Observable from RxJS Subject.
- * @example <caption>Example create a Observable for part of the data Subject.</caption>
- * public readonly myPart = CreateObservablePart(this._data, (data) => data.myPart);
- */
-export function createObservablePart<T, R> (
-	source$: Observable<T>,
-		mappingFunction: MappingFunction<T, R>,
-		memoizationFunction?: MemoizationFunction<R>
-	): Observable<R> {
-	return source$.pipe(
-		map(mappingFunction),
-		distinctUntilChanged(memoizationFunction || defaultMemoization),
-		shareReplay(1)
-	)
-}
-
 
 /**
  * @export
